==> 01_SPE_06_Create_Tidy_DataSet.ipynb <==
!python -V

# iPython Notebook -- Create Tidy DataSet

- Change the location of your {home} dataset before running this notebook
- The structure of files required in this notebook is similar to this:

    {home}:
      |
      --> well_lat_lng.csv (latitude & longitude coordinates)
      |
      --> {SPE_006_originalData/OilSandsDB}
            |
            --> PICKS_DIC.TXT
            --> PICKS.TXT
            --> WELLS.TXT
            --> {Logs}
                  |
                  --> All logs files (las files)
 

## Import Libraries

import os
import glob
import pandas as pd
import numpy as np
import holoviews as hv
import collections
import matplotlib.pyplot as plt
%matplotlib inline
import lasio
lasio.__version__

## Home directory

home = "./SPE_006_originalData/OilSandsDB"
pd.set_option('display.width', 120)

## Read well locations

# read Site ID (SitID) with lat and long coordinates
locations = pd.read_csv("./well_lat_lng.csv", delimiter=',', dtype=str)
locations = locations[['SitID', 'lng', 'lat']]
locations.head()

## Read dictionary of Stratigraphic picks

# describe the stratigraphic 'picks' used in Picks
picks_dicts = pd.read_csv(home + "/PICKS_DIC.TXT", delimiter='\t', dtype=str)
picks_dicts.columns = ['HorID', 'Description']
picks_dicts.head()

## Read stratigraphic picks 

# stratigraphic picks and dropping NaN values
picks = pd.read_csv(home + "/PICKS.TXT", delimiter='\t')
picks['SitID'] = picks['SitID'].astype(str)
picks['HorID'] = picks['HorID'].astype(str)
picks['Pick'] = picks['Pick'].str.strip()
picks['Pick'] = pd.to_numeric(picks['Pick'])
# drop rows with NA values
picks = picks.dropna()
picks.head()

## Read wells with SiteID and UWI

# list of wells
wellst = pd.read_csv(home + "/WELLS.TXT", delimiter= '\t', dtype=str)
wellst.head()

## Merging Wells, Locations, and Picks by SiteID

# merge wellst, locations, and picks in one dataframe
wl = pd.merge(left=wellst, right=locations, left_on=['SitID'], right_on=['SitID'], how='left')
wlp = pd.merge(left=picks, right=wl, left_on=['SitID'], right_on=['SitID'], how='left')
wlp.columns

wlp.head()

## Read LAS files (display number of files read)

# las files
path_logs = home + "/Logs"
# las = [x for x in os.listdir(path_logs + "/") if x.endswith(".LAS")]
# print("# LAS files: " + str(len(las)))

# # a sample of curves from a log file
# wellname = 'AB-07-12-093-10W4-0'
# lasio.read(path_logs + "/" + wellname + ".LAS").df().head()

## Reading actual LAS Files with lasio library

las = [lasio.read(path_logs + "/" + filename) for filename in os.listdir(path_logs) if filename.endswith(".LAS")]
type(las), type(las[0])

## Display some components of lasio datatype

### Header

# las[2153].header['Well']

### Parameters

# las[2153].header['Parameter']

### Version

# las[2153].header['Version']

### Curves available

# las[2153].header['Curves']

### Type of Curves available

# # type of curves (logs) present in the las file
# las[2153].header['Curves'].keys()

### Location of UWI may vary inside the file - Function to retrieve UWI

# function to retrieve UWI (location of the UWI may vary within the LAS file)
keywds = ['UWI', 'UWI:1', 'UWI:2', 'UWI:3']
def getKeyValue(idx, keywds):
    for keywd in keywds:
        if (keywd) not in las[idx].header["Well"]:
            pass
        else:
            return [keywd, las[idx].header["Well"][keywd].value]

### Counting how many curves exists across all wells

# counting more common curves
curves = []
for i in range(len(las)):
    df_las = las[i].df()
    curves.extend(list(df_las.columns))

### Counting and Ordering unique data set of curves

# ordered set of curves across all logs files
counter = collections.Counter(curves)
counter = collections.OrderedDict(sorted(counter.items()))
print(counter.keys())
print(counter.values())

### Selecting common set of curves across all wells

# Selecting only those wells which have common curves: DEPT; DPHI; GR; ILD; NPHI
keys = ['DEPT', 'DPHI', 'DPHI:1', 'DPHI:2', 'GR', 'GR:1', 'GR:2', 'ILD', 'ILD:1', 'ILD:2', \
        'NPHI', 'NPHI:1', 'NPHI:2', 'PHIN', 'PHIN:1', 'PHIN:2']
values = ['DEPT', 'DPHI', 'DPHI', 'DPHI', 'GR', 'GR', 'GR', 'ILD', 'ILD', 'ILD', \
          'NPHI', 'NPHI', 'NPHI', 'NPHI', 'NPHI', 'NPHI']
dictOfCurves = dict(zip(keys, values))

columns = ['i', 'key_uwi', 'uwi', 'curvesNew', 'curvesOld']
rows = []

### Create dataframe with standardized set of curves [GR, DPHI, ILD, DEPT, NPHI]

# create dataframe with standardized set of curves [GR, DPHI, ILD, DEPT, NPHI]
for i in range(len(las)):
    las_curves = las[i].header['Curves'].keys()
    c = [str(dictOfCurves[c]) for c in keys if c in las_curves]
    key_uwi, uwi = getKeyValue(i, keywds)
    rows.append([i, key_uwi, uwi, list(set(c)), las_curves])
    
df = pd.DataFrame(rows, columns=columns)
rows = []
df = df.reset_index(drop=True)

df.head()

len(df)

# drop rows with less than 5 curves including DEPT
dfl = df[df['curvesNew'].map(len) > 4]
dfl = dfl.reset_index(drop=True)

len(dfl)

dfl.head()

# final dataframe with wells having these curves [DEPT, DPHI, GR, ILD, NPHI]
dl = pd.DataFrame([], columns=['UWI', 'DEPT', 'DPHI', 'GR', 'ILD', 'NPHI'])
dl.columns

for i in range(len(dfl)):
    well_num = dfl.iloc[i]['i']
    uwi_type = dfl.iloc[i]['key_uwi']
    uwi = las[well_num].header['Well'][uwi_type].value

    nrecords = len(las[well_num].header['Curves']['DEPT'].data)

    UWI = pd.Series([uwi] * nrecords)
    DEPT = pd.Series(las[well_num].header['Curves']['DEPT'].data)
    DPHI = pd.Series(las[well_num].header['Curves']['DPHI'].data)
    GR = pd.Series(las[well_num].header['Curves']['GR'].data)
    ILD = pd.Series(las[well_num].header['Curves']['ILD'].data)
    NPHI = pd.Series(las[well_num].header['Curves']['NPHI'].data)

    d = {
        'UWI': UWI,
        'DEPT': DEPT,
        'DPHI': DPHI,
        'GR': GR,
        'ILD': ILD,
        'NPHI': NPHI
    }
    dl = dl.append(pd.DataFrame(d), ignore_index=True)

dl.head()

# get UWI (AGS) using a not-well-formed  UWI
def getUWIAGS(wrongUWI):

    s1 = wrongUWI[9:12]
    s2 = wrongUWI[16:17]
    s3 = wrongUWI[13:15]
    s4 = wrongUWI[6:8]
    s5 = wrongUWI[3:5]
    s6 = wrongUWI[0:2]
    s7 = wrongUWI[18:19]
    
    return(s1 + s2 + s3 + s4 + s5 + s6 + s7)

# getUWIAGS('AD/05-15-088-08W400')

# len(dl)

# (dl.iloc[1474750].UWI, getUWIAGS(dl.iloc[1474750].UWI))

# adding UWI_AGS column to join table Wellst which has the SitID. SitID is common to all data tables
dl['UWI_AGS'] = [getUWIAGS(dl.iloc[i].UWI) for i in range(len(dl))]

dl.head()

# see how this 'not-well-formed UWI': AD/05-15-088-08W400, should be this: AD/05-15-088-08W4/0
dl.tail()

### Merging dataframe 'dl' with logs and wells table (wellst) to get SitID

# merging dataframe 'dl' with logs and wells table (wellst) to get SitID  
dl = pd.merge(dl, wellst[['SitID', 'UWI (AGS)']], left_on='UWI_AGS', right_on='UWI (AGS)', how='left')
dt = dl[['SitID', 'DEPT', 'DPHI', 'GR', 'ILD', 'NPHI']]
dt.head()

len(dt)

dt.tail()

dt.index

### Merging dataframe 'dt' with 'wlp' on SitID and [DEPT/Pick]

# merging dataframe 'dt' with 'wlp' on SitID and [DEPT/Pick]
m = pd.merge(left=dt, right=wlp, left_on=['SitID', 'DEPT'], right_on=['SitID', 'Pick'], how='left')
# change HorID to string
m['HorID'] = m['HorID'].astype(str)
# drop NA values
m = m.dropna()
m.head()

len(m)

### Saving Final dataset for ML Modeling

m = m.reset_index(drop=True)
m['Id'] = m.index.values
m = m[['Id', 'SitID', 'DPHI', 'GR', 'ILD', 'NPHI', 'lng', 'lat', 'HorID', 'Pick']]
# m.to_csv('m_out.csv', sep="\t", index=None)

### Adding 5 obervations before & after to every single observation at the DEPT indicated

# reading data file with 5 observation before and after. Prepared in R by Arun Atmacharan
obs_befr_aftr = pd.read_csv("./obs_reduced.csv", delimiter=',')

obs_befr_aftr.head()

obs_befr_aftr.columns.to_series().groupby(obs_befr_aftr.dtypes).groups

obs_befr_aftr.SitIDr = obs_befr_aftr.SitIDr.astype(str)
obs_befr_aftr.columns.to_series().groupby(obs_befr_aftr.dtypes).groups

m.head()

m.columns.to_series().groupby(m.dtypes).groups

# merging dataframe 'm' with 'obs_befr_aftr' on SitID and [DEPT/Pick]
ds = pd.merge(left=m, right=obs_befr_aftr, left_on=['SitID', 'Pick'], right_on=['SitIDr', 'DEPTr'], how='left')

len(ds)

ds = ds.dropna() # to be sure we don't get null values or incomplete observations

len(ds)

ds.head()

ds.columns

ds = ds[['SitID', 'HorID', 'lng', 'lat', \
         'DPHI', 'GR', 'ILD', 'NPHI', \
         'DPHIr1', 'GRr1', 'ILDr1', 'NPHIr1', \
         'DPHIr2', 'GRr2', 'ILDr2', 'NPHIr2', \
         'DPHIr3', 'GRr3', 'ILDr3', 'NPHIr3', \
         'DPHIr4', 'GRr4', 'ILDr4', 'NPHIr4', \
         'DPHIr5', 'GRr5', 'ILDr5', 'NPHIr5', \
         'DPHIr6', 'GRr6', 'ILDr6', 'NPHIr6', \
         'DPHIr7', 'GRr7', 'ILDr7', 'NPHIr7', \
         'DPHIr8', 'GRr8', 'ILDr8', 'NPHIr8', \
         'DPHIr9', 'GRr9', 'ILDr9', 'NPHIr9', \
         'DPHIr10', 'GRr10', 'ILDr10', 'NPHIr10', \
         'Pick']]
ds.head()

# occurrences of every HorID
ds.HorID.value_counts()

# ds['id'] = ds.index
cols = ['id', 'SitID', 'HorID', 'lng', 'lat', 'DPHI', 'GR', 'ILD', 'NPHI',
       'DPHIr1', 'GRr1', 'ILDr1', 'NPHIr1', 'DPHIr2', 'GRr2', 'ILDr2',
       'NPHIr2', 'DPHIr3', 'GRr3', 'ILDr3', 'NPHIr3', 'DPHIr4', 'GRr4',
       'ILDr4', 'NPHIr4', 'DPHIr5', 'GRr5', 'ILDr5', 'NPHIr5', 'DPHIr6',
       'GRr6', 'ILDr6', 'NPHIr6', 'DPHIr7', 'GRr7', 'ILDr7', 'NPHIr7',
       'DPHIr8', 'GRr8', 'ILDr8', 'NPHIr8', 'DPHIr9', 'GRr9', 'ILDr9',
       'NPHIr9', 'DPHIr10', 'GRr10', 'ILDr10', 'NPHIr10', 'Pick']
ds = ds[['id', 'SitID', 'HorID', 'lng', 'lat', 'DPHI', 'GR', 'ILD', 'NPHI',
       'DPHIr1', 'GRr1', 'ILDr1', 'NPHIr1', 'DPHIr2', 'GRr2', 'ILDr2',
       'NPHIr2', 'DPHIr3', 'GRr3', 'ILDr3', 'NPHIr3', 'DPHIr4', 'GRr4',
       'ILDr4', 'NPHIr4', 'DPHIr5', 'GRr5', 'ILDr5', 'NPHIr5', 'DPHIr6',
       'GRr6', 'ILDr6', 'NPHIr6', 'DPHIr7', 'GRr7', 'ILDr7', 'NPHIr7',
       'DPHIr8', 'GRr8', 'ILDr8', 'NPHIr8', 'DPHIr9', 'GRr9', 'ILDr9',
       'NPHIr9', 'DPHIr10', 'GRr10', 'ILDr10', 'NPHIr10', 'Pick']]

ds.head()

# writing final dataset to disk - ready for NIA
ds.to_csv('ds_out_ready.csv', sep="\t", index=None)

### end of Create tidy dataset
