==> 02_SPE_06_ModelFitting_v0.1.ipynb <==
# Install a conda package in the current Jupyter kernel
import sys
# !conda install --yes --prefix {sys.prefix} -c conda-forge auto-sklearn
!conda install --yes --prefix {sys.prefix} -c conda-forge pyrfr

import os
import pandas as pd
import numpy as np
import holoviews as hv
import collections
import matplotlib.pyplot as plt
%matplotlib inline

import sklearn.model_selection
import sklearn.datasets
import sklearn.metrics

import autosklearn.regression

### Read dataset

ds = pd.read_csv("./ds_out_ready.csv", delimiter='\t', dtype=str)

ds.columns

ds = ds[['Pick', 'HorID', 'lng', 'lat', 'DPHI', 'GR', 'ILD', 'NPHI',
       'DPHIr1', 'GRr1', 'ILDr1', 'NPHIr1', 'DPHIr2', 'GRr2', 'ILDr2',
       'NPHIr2', 'DPHIr3', 'GRr3', 'ILDr3', 'NPHIr3', 'DPHIr4', 'GRr4',
       'ILDr4', 'NPHIr4', 'DPHIr5', 'GRr5', 'ILDr5', 'NPHIr5', 'DPHIr6',
       'GRr6', 'ILDr6', 'NPHIr6', 'DPHIr7', 'GRr7', 'ILDr7', 'NPHIr7',
       'DPHIr8', 'GRr8', 'ILDr8', 'NPHIr8', 'DPHIr9', 'GRr9', 'ILDr9',
       'NPHIr9', 'DPHIr10', 'GRr10', 'ILDr10', 'NPHIr10']]

ds.head()
