==> 03_SPE_06_OneHotEncodingQuality.ipynb <==
library('tidyverse')
library("stringr")

print(getwd())

data <- read_csv("10aroundPlus3Near.csv")

head(data)

to_Colnames <- colnames(data)
colnames(data) <- stringr::str_replace_all(string = to_Colnames, pattern = "\\.", replacement = "_")
colnames(data)

length(data$id)

unique(data$Quality)

# exclude variables 'SitID', 'Pick', 'UWI (AGS)', 'UWI', 'Picked'
theseVars <- names(data) %in% c('SitID', 'Pick', 'UWI (AGS)', 'UWI', 'Picked') 
data <- data[!theseVars]

df <- data[, c('id', 'Quality')]

# onehot encoding is a representation of categorical variables as binary vectors. 
# What this means is that we want to transform a categorical variable or variables 
# to a format that works better with classification and regression algorithms.
for(unique_value in unique(df$Quality)) {
    df[paste("Quality", unique_value, sep = "_")] <- ifelse(df$Quality == unique_value, 1, 0)
}
df$Quality <- NULL

dfmerged <- merge(x = data, y = df, by = "id", all.x = TRUE)
dfmerged$Quality <- NULL

dfmerged$HorID <- as.character(dfmerged$HorID)
dfmerged$Quality_1 <- as.character(dfmerged$Quality_1)
dfmerged$Quality_2 <- as.character(dfmerged$Quality_2)
dfmerged$Quality_3 <- as.character(dfmerged$Quality_3)
dfmerged$Quality_4 <- as.character(dfmerged$Quality_4)

str(dfmerged)

# write.csv(file = "onehotEncodedQualityDSet.csv", x = dfmerged, 
#           quote = c(7, 62, 63, 64, 65), row.names = FALSE)
write.table(x = dfmerged, file = "onehotEncodedQualityDSet.csv", 
            row.names = FALSE, sep="\t")
