==> 04_SPE_06_Creating New DataSet and Model XGBoostTree.ipynb <==
getwd()

## 1.0 Import Libraries

library(tidyverse)
library(stringr)
library(ggplot2)
library(RMySQL)

## 2.0 Read Data

### 2.1 Connect database

# set credentials
user <- "root"
password <- "Dario5793"
database_name <- "dblas"
host <- "localhost"

# open connection
con = dbConnect(MySQL(), user=user, password=password, dbname=database_name, host=host)

### 2.2 Create generic SQL Statements

# sql statements
statement_well <- "select SitID, HorID, Pick, lat, lng from well_lat_lng;"
statement_logs <- "select * from logs limit 10;"

### 2.3 Read data

### 2.3.1 Well data

# create sql statement
statement <- "select SitID, HorID, Pick, lat, lng from well_lat_lng;"
rs <- dbSendQuery(con, statement)
# fetch data from result set and convert to a tibble
well <- as_tibble(fetch(rs, n = 20))
class(well)
hasCompltd <- dbHasCompleted(rs)
isCleardRS <- dbClearResult(rs)

head(well)

### 2.3.2 Logs data

# create sql statement
statement <- "select * from logs limit 20;"
rs <- dbSendQuery(con, statement)
# fetch data from result set and convert to a tibble
logs <- as_tibble(fetch(rs, n = 20))
class(logs)
hasCompltd <- dbHasCompleted(rs)
isCleardRS <- dbClearResult(rs)

head(logs)

require(purrr)

row_num <- 7
# data <- as.matrix(filter(logs, between(row_number(), row_num + 1, row_num + row_num - 1)))
data <- as.matrix(filter(logs, between(row_number(), row_num + 1, row_num + 5)))
data_aft <- as_tibble(cbind(nms = names(c('SitID', 'DEPT', 'DPHI', 'GR', 'ILD', 'NPHI')), t(data)))

data <- as.matrix(filter(logs, between(row_number(), -row_num, row_num)))
data_bef <- as_tibble(cbind(nms = names(c('SitID', 'DEPT', 'DPHI', 'GR', 'ILD', 'NPHI')), t(data)))

data_bef

data

x = tibble(value = c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20))
# x
row_righ = 7
row_left = row_righ - 5
t(filter(x, between(row_number(), left = row_left, right = row_righ - 1)))

row_left = 7
row_righ = row_left + 5
t(filter(x, between(row_number(), left = row_left + 1, right = row_righ)))

dbDisconnect(con)

## 3.0 Create Model

## 4.0 Train Model

## 5.0 Test Model
