==> 05_SPE_06_kNN_Mannville_Wells.ipynb <==
library(tidyverse)
library(purrr)

t1 <- read_csv(file = 'dataset_distinct_lat_lng_SitID_incl.csv')
t2 <- read_csv(file = 'dataset_result_nn.csv')
# join t1 and t2
df <- left_join(t1, t2, by = 'SitID')
# add column with row number
df$num_row <- 1:nrow(df)

head(df)

tail(df)

write_csv(x = df, path = 'dataset_kNN_Wells_Mannville.csv')

knn <- read_csv(file = './neighbors/neighbors_SitID_lng_lat.csv')

head(knn, 10)
