==> 07_SPE_06_DNN_with_Tensorflow_Modeling.ipynb <==
# 1.0 Load libraries

import tensorflow as tf
import pandas as pd
import numpy as np
from scipy.stats import zscore

from pandas import read_csv 
from keras.models import Sequential 
from keras.layers import Dense 
from keras.wrappers.scikit_learn import KerasClassifier 
from keras.utils import np_utils 
from sklearn.model_selection import cross_val_score 
from sklearn.model_selection import KFold 
from sklearn.preprocessing import LabelEncoder 
from sklearn.pipeline import Pipeline

# Read data

# reading dataset created by Dario
df = pd.read_csv('dataset_normalized_20180221_1637.csv')

pd.set_option('display.width', 340)

# Create a Model - hyperparameters

# Parameters
features = [x for x in df.columns if x != "HorID"]
num_features = len(features)
num_steps = 500 # Total steps to train
batch_size = 1024 # The number of samples per batch
classes = df['HorID'].unique()
num_classes = len(classes) # The 14 HorID's
num_trees = 10
max_nodes = 1000

# Only two classes 5000 & 14000 for a binary classification
df_binary = df.loc[df['HorID'].isin([5000, 14000])]
df_binary.HorID = df_binary.HorID.astype(str, inplace=True)
# df_binary.dtypes

horizon = tf.feature_column.categorical_column_with_vocabulary_list('horizon', ['5000', '14000'])

dept = tf.feature_column.numeric_column('DEPT')
reading = tf.feature_column.numeric_column('reading')
lng = tf.feature_column.numeric_column('lng')
lat = tf.feature_column.numeric_column('lat')
cType_DPHI = tf.feature_column.numeric_column('cType_DPHI')
cType_GR = tf.feature_column.numeric_column('cType_GR')
cType_ILD = tf.feature_column.numeric_column('cType_ILD')
cType_NPHI = tf.feature_column.numeric_column('cType_NPHI')
cType_DPHI1 = tf.feature_column.numeric_column('cType_DPHI1')
cType_GR1 = tf.feature_column.numeric_column('cType_GR1')
cType_ILD1 = tf.feature_column.numeric_column('cType_ILD1')
cType_NPHI1 = tf.feature_column.numeric_column('cType_NPHI1')
cType_DPHI2 = tf.feature_column.numeric_column('cType_DPHI2')
cType_GR2 = tf.feature_column.numeric_column('cType_GR2')
cType_ILD2 = tf.feature_column.numeric_column('cType_ILD2')
cType_NPHI2 = tf.feature_column.numeric_column('cType_NPHI2')
cType_DPHI3 = tf.feature_column.numeric_column('cType_DPHI3')
cType_GR3 = tf.feature_column.numeric_column('cType_GR3')
cType_ILD3 = tf.feature_column.numeric_column('cType_ILD3')
cType_NPHI3 = tf.feature_column.numeric_column('cType_NPHI3')
cType_DPHI4 = tf.feature_column.numeric_column('cType_DPHI4')
cType_GR4 = tf.feature_column.numeric_column('cType_GR4')
cType_ILD4 = tf.feature_column.numeric_column('cType_ILD4')
cType_NPHI4 = tf.feature_column.numeric_column('cType_NPHI4')
cType_DPHI5 = tf.feature_column.numeric_column('cType_DPHI5')
cType_GR5 = tf.feature_column.numeric_column('cType_GR5')
cType_ILD5 = tf.feature_column.numeric_column('cType_ILD5')
cType_NPHI5 = tf.feature_column.numeric_column('cType_NPHI5')
cType_DPHI6 = tf.feature_column.numeric_column('cType_DPHI6')
cType_GR6 = tf.feature_column.numeric_column('cType_GR6')
cType_ILD6 = tf.feature_column.numeric_column('cType_ILD6')
cType_NPHI6 = tf.feature_column.numeric_column('cType_NPHI6')
cType_DPHI7 = tf.feature_column.numeric_column('cType_DPHI7')
cType_GR7 = tf.feature_column.numeric_column('cType_GR7')
cType_ILD7 = tf.feature_column.numeric_column('cType_ILD7')
cType_NPHI7 = tf.feature_column.numeric_column('cType_NPHI7')
cType_DPHI8 = tf.feature_column.numeric_column('cType_DPHI8')
cType_GR8 = tf.feature_column.numeric_column('cType_GR8')
cType_ILD8 = tf.feature_column.numeric_column('cType_ILD8')
cType_NPHI8 = tf.feature_column.numeric_column('cType_NPHI8')
cType_DPHI9 = tf.feature_column.numeric_column('cType_DPHI9')
cType_GR9 = tf.feature_column.numeric_column('cType_GR9')
cType_ILD9 = tf.feature_column.numeric_column('cType_ILD9')
cType_NPHI9 = tf.feature_column.numeric_column('cType_NPHI9')
cType_DPHI10 = tf.feature_column.numeric_column('cType_DPHI10')
cType_GR10 = tf.feature_column.numeric_column('cType_GR10')
cType_ILD10 = tf.feature_column.numeric_column('cType_ILD10')
cType_NPHI10 = tf.feature_column.numeric_column('cType_NPHI10')
cType_DPHI_near_1 = tf.feature_column.numeric_column('cType_DPHI_near_1')
cType_GR_near_1 = tf.feature_column.numeric_column('cType_GR_near_1')
cType_ILD_near_1 = tf.feature_column.numeric_column('cType_ILD_near_1')
cType_NPHI_near_1 = tf.feature_column.numeric_column('cType_NPHI_near_1')
cType_DPHI_near_2 = tf.feature_column.numeric_column('cType_DPHI_near_2')
cType_GR_near_2 = tf.feature_column.numeric_column('cType_GR_near_2')
cType_ILD_near_2 = tf.feature_column.numeric_column('cType_ILD_near_2')
cType_NPHI_near_2 = tf.feature_column.numeric_column('cType_NPHI_near_2')
cType_DPHI_near_3 = tf.feature_column.numeric_column('cType_DPHI_near_3')
cType_GR_near_3 = tf.feature_column.numeric_column('cType_GR_near_3')
cType_ILD_near_3 = tf.feature_column.numeric_column('cType_ILD_near_3')
cType_NPHI_near_3 = tf.feature_column.numeric_column('cType_NPHI_near_3')
