==> 08_SPE_06_DNN_with_Tensorflow_R_02.ipynb <==
library(tfestimators)
library(tensorflow)
library(tidyverse)
library(operators)

getwd()

setwd("./dataset_FINAL_20180225_1556_for_Modeling.csv")

# read data
strata <- read_csv("dataset_FINAL_20180225_1556_for_Modeling.csv")
head(strata, 6)

# factor-ize HorID column
strata$HorID <- factor(strata$HorID)

response <- "DEPT"
features <- setdiff(names(strata), response)

# split into train, test datasets
set.seed(123)
partitions <- modelr::resample_partition(strata, c(test = 0.15, train = 0.85))
strata_train <- as.data.frame(partitions$train)
strata_test  <- as.data.frame(partitions$test)

head(strata_train)

# construct regressor
regressor <- dnn_regressor(
  feature_columns = feature_columns, 
  hidden_units = c(20, 40, 20),
  # hidden_units = c(6, 10, 20, 10, 4), 
  n_classes = length(unique(strata$HorID)), 
  activation_fn = "relu", optimizer = "Adagrad"
)

# construct input function 
strata_input_fn <- function(data) {
  input_fn(data, features = features(), response = response())
}

# train classifier with training dataset
train(classifier, input_fn = strata_input_fn(strata_train), max_steps = 10000)

# valuate with test dataset
predictions <- predict(classifier, input_fn = strata_input_fn(strata_test))
evaluation <- evaluate(classifier, input_fn = strata_input_fn(strata_test))

evaluation
