==> 10_SPE_06_get_DS_Tidy_with_Neighbors.ipynb <==
# concatenate results for 20 knn of interest per well

# read file 'left' of the join
file_path <- "../neighbors/"
file_extn <- ".csv"
file_name <- "neighbors_10aroundPlus3Near_20toModelingNoSitID"

# subset of data of interest
ds_left <- get_dataset(file_path, file_name, file_extn)

# read file 'right' of the join
file_path <- "../"
file_extn <- ".csv"
file_name <- "10aroundPlus3Near_toModeling"

# subset of data of interest
ds_right <- get_dataset(file_path, file_name, file_extn) %>% 
  select(SitID, HorID, lng, lat) %>% filter(complete.cases(.))

# left join on SitID using ds_left and ds_right dataframes 
ds <- dplyr::left_join(x = ds_left, y = ds_right, by = 'SitID')

# show column names
colnames(ds)

# initialize variables
col_shift <- 1 # skip first 4 columns
c_col <- 1 # column for SitID
n_rows <- nrow(ds)

# create 10 new variables of neighbors SitID's around each SitID location
for (i in seq(1:n_rows)) {
    for (j in seq(1:20)) {
        c_row <- as.integer(ds[i, col_shift + j] + 1)
        ds[i, paste0("SitID", j, sep = "")] <- 
            assign(paste0("SitID", j, sep = ""), ds[c_row, 'SitID'])
    }
    print(paste0('row [', i, '] out of ', n_rows))
}

# write resultant dataframe
print(paste('writing file', paste0(file_path, "/", file_name, "_extnd", file_extn))) 
write.csv(x = ds, file = paste0(file_path, "/", "ds", "_extnd", file_extn), 
          row.names = FALSE)
