==> 12_SPE_06_DataWrangling_FINAL.ipynb <==
# import tensorflow as tf
import pandas as pd
import numpy as np
from scipy.stats import zscore

from pandas import read_csv 
# from keras.models import Sequential 
# from keras.layers import Dense 
# from keras.wrappers.scikit_learn import KerasClassifier 
from keras.utils import np_utils 
from sklearn.model_selection import cross_val_score 
from sklearn.model_selection import KFold 
from sklearn.preprocessing import LabelEncoder 
from sklearn.pipeline import Pipeline

# print(tf.__version__)

# Install a pip package in the current Jupyter kernel
# import sys
# !{sys.executable} -m pip install scipy --user

# reading dataset created by Arun
df = pd.read_csv("dataset_10aroundPlus3Near_FINAL.csv")
df.columns

df.tail()

df_columns = ['UWI', 'SitID', 'DEPT', 'HorID', 'lng', 'lat', 
              'DPHI', 'GR', 'ILD', 'NPHI', 
              'DPHI1', 'GR1', 'ILD1', 'NPHI1', 'DPHI2', 'GR2', 'ILD2', 'NPHI2', 
              'DPHI3', 'GR3', 'ILD3', 'NPHI3', 'DPHI4', 'GR4', 'ILD4', 'NPHI4', 
              'DPHI5', 'GR5', 'ILD5', 'NPHI5', 'DPHI6', 'GR6', 'ILD6', 'NPHI6', 
              'DPHI7', 'GR7', 'ILD7', 'NPHI7', 'DPHI8', 'GR8', 'ILD8', 'NPHI8', 
              'DPHI9', 'GR9', 'ILD9', 'NPHI9', 'DPHI10', 'GR10', 'ILD10', 'NPHI10', 
              'DPHI_ner1', 'GR_ner1', 'ILD_ner1', 'NPHI_ner1', 
              'DPHI_ner2', 'GR_ner2', 'ILD_ner2', 'NPHI_ner2', 
              'DPHI_ner3', 'GR_ner3', 'ILD_ner3', 'NPHI_ner3']

# drop 'id' column from dataset
df.drop('id', axis=1, inplace=True)

df.reset_index(inplace=True, drop=True)
df.shape

df = df[df_columns].copy(deep=True)
df.tail()

len(df['UWI'].unique())

# generate a random sample of 3 integers as indexes for the prediction case
from random import sample
n_uwis = len(df['UWI'].unique())
list_for_prediction = sample(range(n_uwis), 3)
print(type(list_for_prediction), list_for_prediction)

df = df.iloc[~(df.index.isin(list_for_prediction))]
df_pred = df.iloc[list_for_prediction]

final_columns = ['DEPT', 'HorID', 'lng', 'lat', 'DPHI', 'GR', 'ILD',
       'NPHI', 'DPHI1', 'GR1', 'ILD1', 'NPHI1', 'DPHI2', 'GR2', 'ILD2',
       'NPHI2', 'DPHI3', 'GR3', 'ILD3', 'NPHI3', 'DPHI4', 'GR4', 'ILD4',
       'NPHI4', 'DPHI5', 'GR5', 'ILD5', 'NPHI5', 'DPHI6', 'GR6', 'ILD6',
       'NPHI6', 'DPHI7', 'GR7', 'ILD7', 'NPHI7', 'DPHI8', 'GR8', 'ILD8',
       'NPHI8', 'DPHI9', 'GR9', 'ILD9', 'NPHI9', 'DPHI10', 'GR10', 'ILD10',
       'NPHI10', 'DPHI_ner1', 'GR_ner1', 'ILD_ner1', 'NPHI_ner1', 'DPHI_ner2',
       'GR_ner2', 'ILD_ner2', 'NPHI_ner2', 'DPHI_ner3', 'GR_ner3', 'ILD_ner3',
       'NPHI_ner3']

df_pred.head()

df.head()

## FINAL dataset for Modeling

final_columns = ['DEPT', 'HorID', 'lng', 'lat', 'DPHI', 'GR', 'ILD',
       'NPHI', 'DPHI1', 'GR1', 'ILD1', 'NPHI1', 'DPHI2', 'GR2', 'ILD2',
       'NPHI2', 'DPHI3', 'GR3', 'ILD3', 'NPHI3', 'DPHI4', 'GR4', 'ILD4',
       'NPHI4', 'DPHI5', 'GR5', 'ILD5', 'NPHI5', 'DPHI6', 'GR6', 'ILD6',
       'NPHI6', 'DPHI7', 'GR7', 'ILD7', 'NPHI7', 'DPHI8', 'GR8', 'ILD8',
       'NPHI8', 'DPHI9', 'GR9', 'ILD9', 'NPHI9', 'DPHI10', 'GR10', 'ILD10',
       'NPHI10', 'DPHI_ner1', 'GR_ner1', 'ILD_ner1', 'NPHI_ner1', 'DPHI_ner2',
       'GR_ner2', 'ILD_ner2', 'NPHI_ner2', 'DPHI_ner3', 'GR_ner3', 'ILD_ner3',
       'NPHI_ner3']

df = df[final_columns].copy(deep=True)

df.to_csv(path_or_buf='dataset_FINAL_20180225_1556_for_Modeling.csv', index=False)
df_pred.to_csv(path_or_buf='dataset_FINAL_20180225_1556_for_Predict.csv', index=False)

dz = pd.DataFrame(np.random.randint(0, 10, (4,4)), columns=list('ABCD'))
dz.head()

dz.iloc[~(dz.index.isin([1, 2]))]

# create unique set of well locations - will be used later as a identifiers
df_uni = (df[['lng', 'lat']].drop_duplicates().reset_index())[['lng', 'lat']]
df_uni['well_loc'] = df_uni.index.values
df = pd.merge(df_uni, df, how='left', on=['lng', 'lat'])

# show a sample of the 'horizontal' observations
df.reset_index()
df[['well_loc', 'lng', 'lat', 'DEPT', 'DPHI', 'GR', 'ILD', 'NPHI', 'HorID', 'DPHI1', 'NPHI_near_3']].tail()

df[['well_loc', 'lng', 'lat', 'DEPT', 'DPHI', 'GR', 'ILD', 'NPHI', 'HorID', 'DPHI1', 'NPHI_near_3']].head()

df.shape

# calculate zscore for every column in a dataframe
def make_zscore(columns = [], df = pd.DataFrame()):
    ds = pd.DataFrame()
    for col in columns:
        ds[col] = df[col].to_frame().apply(zscore, ddof=1)
    return(ds)

def verticalize_ds(anchor_col = 'DEPT', value = 'reading', ds = pd.DataFrame()):
    # verticalize dataset by log observation
    data = pd.DataFrame()
    data = (ds.replace(0, np.nan)
              .set_index(anchor_col, append=True)
              .stack()
              .reset_index()
              .rename(columns={0:value}))
    data.columns = ['p_idx', anchor_col, 'CurveType', value]
    return(data)

## Create Z-Score Values - dataset 'ds'

# create zscore values for numerical log columns
# log_columns = ['DEPT', 'DPHI', 'GR', 'ILD', 'NPHI', 
#            'DPHI1', 'GR1', 'ILD1', 'NPHI1', 'DPHI2', 'GR2', 'ILD2', 'NPHI2', 
#            'DPHI3', 'GR3', 'ILD3', 'NPHI3', 'DPHI4', 'GR4', 'ILD4', 'NPHI4', 
#            'DPHI5', 'GR5', 'ILD5', 'NPHI5', 'DPHI6', 'GR6', 'ILD6', 'NPHI6', 
#            'DPHI7', 'GR7', 'ILD7', 'NPHI7', 'DPHI8', 'GR8', 'ILD8', 'NPHI8', 
#            'DPHI9', 'GR9', 'ILD9', 'NPHI9', 'DPHI10', 'GR10', 'ILD10', 'NPHI10', 
#            'DPHI_near_1', 'GR_near_1', 'ILD_near_1', 'NPHI_near_1', 
#            'DPHI_near_2', 'GR_near_2', 'ILD_near_2', 'NPHI_near_2', 
#            'DPHI_near_3', 'GR_near_3', 'ILD_near_3', 'NPHI_near_3']
log_columns = ['DPHI', 'GR', 'ILD', 'NPHI', 
           'DPHI1', 'GR1', 'ILD1', 'NPHI1', 'DPHI2', 'GR2', 'ILD2', 'NPHI2', 
           'DPHI3', 'GR3', 'ILD3', 'NPHI3', 'DPHI4', 'GR4', 'ILD4', 'NPHI4', 
           'DPHI5', 'GR5', 'ILD5', 'NPHI5', 'DPHI6', 'GR6', 'ILD6', 'NPHI6', 
           'DPHI7', 'GR7', 'ILD7', 'NPHI7', 'DPHI8', 'GR8', 'ILD8', 'NPHI8', 
           'DPHI9', 'GR9', 'ILD9', 'NPHI9', 'DPHI10', 'GR10', 'ILD10', 'NPHI10', 
           'DPHI_near_1', 'GR_near_1', 'ILD_near_1', 'NPHI_near_1', 
           'DPHI_near_2', 'GR_near_2', 'ILD_near_2', 'NPHI_near_2', 
           'DPHI_near_3', 'GR_near_3', 'ILD_near_3', 'NPHI_near_3']
ds = make_zscore(log_columns, df)
ds.tail()

## Merge on a left join based on indexes - dataset _'dt'_

# merge left join on indexes
dt = pd.merge(df[['well_loc', 'lng', 'lat', 'HorID', 'DEPT']], ds, how='left', left_index=True, right_index=True)
dt[['well_loc', 'lng', 'lat', 'DPHI', 'GR', 'ILD', 'NPHI', 'HorID', 'DPHI1', 'NPHI_near_3']].tail()

dt.shape

dt.columns

dt['p_idx'] = dt.index.values
dt_columns = ['p_idx', 'well_loc', 'lng', 'lat', 'HorID', 'DEPT', 'DPHI', \
              'GR', 'ILD', 'NPHI', 'HorID', 'DPHI1', 'NPHI_near_3']
dt[dt_columns].tail()

dt[dt_columns].head()

# verticalize logs data stored on variable 'columns'
data = verticalize_ds('DEPT', 'reading', dt[log_columns])
data.tail()

# left outer join to HorID and lat and long coordinates on DEPT
df_out = pd.merge(data, dt, how='left', on=['p_idx'])
df_out.rename(index=str, columns={"DEPT_x": "DEPT"}, inplace=True)
df_out.drop(['p_idx'], inplace=True, axis=1)
df_out.tail()

df_out.reset_index(drop=True, inplace=True)
# df_out.index.values
df_out.shape

# One Hot Encoding HorID
Y = df_out['CurveType']
Y.head()
# encode class values as integers 
encoder = LabelEncoder() 
encoder.fit(Y) 
encoded_Y = encoder.transform(Y) 
# convert integers to dummy variables (i.e. one hot encoded) 
dummy_y = np_utils.to_categorical(encoded_Y)
dummy_y.shape

curves = df_out['CurveType'].unique()
curves

dummy_cols = ['cType_' + x for x in curves]
df_out_dummy = pd.DataFrame(dummy_y, columns=dummy_cols)
df_out_dummy.columns

df_out = pd.merge(df_out, df_out_dummy, how='left', left_index=True, right_index=True).copy(deep=True)

# pd.set_option('display.width', 240)
cols_to_drop = ['CurveType', 'well_loc', 'DEPT_y','DPHI', 'GR', 'ILD', 'NPHI', 'DPHI1', 'GR1', 'ILD1', 'NPHI1', 'DPHI2', 'GR2', 'ILD2', 'NPHI2', 'DPHI3', 'GR3', 'ILD3', 'NPHI3', 'DPHI4', 'GR4', 'ILD4',
       'NPHI4', 'DPHI5', 'GR5', 'ILD5', 'NPHI5', 'DPHI6', 'GR6', 'ILD6', 'NPHI6', 'DPHI7', 'GR7', 'ILD7', 'NPHI7', 'DPHI8', 'GR8', 'ILD8', 'NPHI8', 'DPHI9', 'GR9', 'ILD9', 'NPHI9', 'DPHI10', 'GR10', 'ILD10', 'NPHI10', 'DPHI_near_1',
       'GR_near_1', 'ILD_near_1', 'NPHI_near_1', 'DPHI_near_2', 'GR_near_2', 'ILD_near_2', 'NPHI_near_2', 'DPHI_near_3', 'GR_near_3', 'ILD_near_3', 'NPHI_near_3']
df_out.drop(cols_to_drop, inplace=True, axis=1)
df_out.head()

df_out[['lng', 'lat']] = make_zscore(['lng', 'lat'], df_out[['lng', 'lat']])

df_out.tail()

df_out.HorID.unique()

mask = df_out['HorID'].isin([1000, 5000,  7000, 11000, 12000, 13000, 14000, 
                             9000,  9500, 10000])
df_out = df_out[mask].copy(deep=True)
df_out.tail()

df_out.to_csv(path_or_buf='dataset_normalized_20180222_0241.csv', index=False)
